Extract Tier-1 provisioning to config-free pkg/provision leaf package
Move ProvisionShared, its helpers, and the ProvisionInput/ResolvedWorkspace/
ResolvedSharedDir types from pkg/runtime to a new pkg/provision package
that depends only on stdlib + pkg/api + pkg/store (no pkg/config).

This fixes TestInitProjectDataIsolation: the lean sciontool binary can now
import pkg/provision for its `provision` subcommand without transitively
pulling in pkg/config (filesystem-based project path resolution).

Backward-compatible type aliases and a thin ProvisionShared wrapper remain
in pkg/runtime so existing callers (nfsBackend, broker) keep compiling.

Tests for moved symbols (sanitizeBranchName, writeSentinel, validation,
acquireProvisionLock context cancellation) moved to pkg/provision.
Integration tests using nfsTestBackend stay in pkg/runtime.
